rename DesignableBannerCtasV2 to DesignableBannerCtas and fix import
